+++ src/app/pages/tabs/analysis/analysis.page.ts
@@ -88,7 +88,6 @@ export class AnalysisPage implements OnDestroy {
             next: (report) => {
               // If there is no active report, it creates a new one and start the analysis.
               if (report === null) {
-                console.log(report);
                 this.continueWithNewReport();
                 return;
               }
@@ -140,7 +139,6 @@ export class AnalysisPage implements OnDestroy {
             next: (report) => {
               // If there is no active report, it creates a new one and start the analysis.
               if (report !== null) {
-                console.log("reporte:", report);
                 this.utilsSvc.routerLink("/tabs/analysis/analysis-form");
                 return;
               }