+++ src/assets/data/help-slides.ts
@@ -2,8 +2,8 @@ export const helpSlides = [
   {
     title: "¡Comencemos!",
     subtitle: "¡Nos encargáremos de analizar la madera por ti!",
-    indication: "Antes que nada crea tu cuenta",
-    icon: "",
+    indication: "",
+    img: "assets/icon/wt_Group465.svg",
     message:
       "Ten en cuenta que para realizar los análisis de madera se requiere tener una membresía activa",
   },
@@ -11,7 +11,7 @@ export const helpSlides = [
     title: "Fotografía de madera",
     subtitle: "Utiliza una lupa con un aumento mínimo de x60",
     indication: "",
-    icon: "",
+    img: "assets/icon/wt_start_help2.svg",
     message:
       "Recuerda que el análisis de la madera depende del uso de esta lupa",
   },
@@ -20,14 +20,14 @@ export const helpSlides = [
     subtitle:
       "Con una sola foto nuestra inteligencia artifical se encargará del análisis",
     indication: "",
-    icon: "",
+    img: "assets/icon/wt_start_help3.svg",
     message: "Podrás identificar y ratificar las maderas siendo transportadas",
   },
   {
     title: "Reportes",
     subtitle: "Crear y guarda los reportes de manera rápida y segura",
     indication: "",
-    icon: "",
-    message: "Puedes descargar los reportes en cualquier momento",
+    img: "assets/icon/wt_start_help4.svg",
+    message: "Puedes descargar los reportes en cualquier momento.",
   },
 ];